Added simple controller for input and drawing.
Used something like this before, where controllers
are on a stack, with the topmost being the active
controller.  The active controller handles the
input, while at the discretion of the active controller,
some inactive controllers can render to the screen
before the active controller does.  (for example,
if you have a controller for a dialog, you still want
the previous controllers to draw, so the screen outside
the dialog doesn't just go blank).